repoとrepoImplの分離
