+++ atcoder_helper/application/repositories/logged_in_session_repo.py
@@ -2,12 +2,48 @@
 
 import os
 import pickle
+from typing import Protocol
 from typing import cast
 
 import requests
 
 from atcoder_helper.application.repositories.errors import ReadError
 from atcoder_helper.application.repositories.errors import WriteError
+
+
+class LoggedInSessionRepository(Protocol):
+    """Login済みセッションのリポジトリのプロトコル."""
+
+    def write(self, session: requests.Session) -> None:
+        """_write.
+
+        Raises:
+            WriteError: 書き込みに失敗
+        """
+
+    def read(self) -> requests.Session:
+        """read. ファイルが存在しない場合はデフォルトsessionを返す.
+
+        Raises:
+            ReadError: 読み込みに失敗した
+
+        Return:
+            requests.Session: 取得したsession
+        """
+
+    def doesExist(self) -> bool:
+        """sessionがすでに存在するかどうか確認する.
+
+        Returns:
+            bool: 存在するかどうか
+        """
+
+    def delete(self) -> None:
+        """Session 情報を削除します.
+
+        Raises:
+            WriteError: _
+        """
 
 
 class LoggedInSessionRepositoryImpl:
