+++ atcoder_helper/dependency.py
@@ -21,6 +21,25 @@ from atcoder_helper.application.interactors.init_task import InitTaskDirInteract
 from atcoder_helper.application.interactors.util import (
     get_atcoder_helper_config_filepath,
 )
+from atcoder_helper.application.repositories.atcoder_helper_config_repo import (
+    ConfigRepository,
+)
+from atcoder_helper.application.repositories.atcoder_logged_in_session_repo import (
+    AtCoderLoggedInSessionRepository,
+)
+from atcoder_helper.application.repositories.atcoder_test_case_repo import (
+    AtCoderTestCaseRepository,
+)
+from atcoder_helper.application.repositories.local_test_case_repo import (
+    LocalTestCaseRepository,
+)
+from atcoder_helper.application.repositories.logged_in_session_repo import (
+    LoggedInSessionRepository,
+)
+from atcoder_helper.application.repositories.login_status_repo import LoginStatusRepo
+from atcoder_helper.application.repositories.task_config_repo import (
+    TaskConfigRepository,
+)
 from atcoder_helper.application.usecases.atcoder_helper_config import (
     AtCoderHelperConfigUsecase,
 )
@@ -28,35 +47,22 @@ from atcoder_helper.application.usecases.auth import AuthUsecase
 from atcoder_helper.application.usecases.execute_test import ExecuteTestUsecase
 from atcoder_helper.application.usecases.fetch_task import FetchTaskUsecase
 from atcoder_helper.application.usecases.init_task import InitTaskDirUsecase
-from atcoder_helper.infrastructure.atcoder_helper_config_repo import ConfigRepository
 from atcoder_helper.infrastructure.atcoder_helper_config_repo import (
     ConfigRepositoryImpl,
-)
-from atcoder_helper.infrastructure.atcoder_logged_in_session_repo import (
-    AtCoderLoggedInSessionRepository,
 )
 from atcoder_helper.infrastructure.atcoder_logged_in_session_repo import (
     AtCoderLoggedInSessionRepositoryImpl,
 )
 from atcoder_helper.infrastructure.atcoder_test_case_repo import (
-    AtCoderTestCaseRepository,
-)
-from atcoder_helper.infrastructure.atcoder_test_case_repo import (
     AtCoderTestCaseRepositoryImpl,
 )
-from atcoder_helper.infrastructure.local_test_case_repo import LocalTestCaseRepository
 from atcoder_helper.infrastructure.local_test_case_repo import (
     LocalTestCaseRepositoryImpl,
 )
 from atcoder_helper.infrastructure.logged_in_session_repo import (
-    LoggedInSessionRepository,
-)
-from atcoder_helper.infrastructure.logged_in_session_repo import (
     LoggedInSessionRepositoryImpl,
 )
-from atcoder_helper.infrastructure.login_status_repo import LoginStatusRepo
 from atcoder_helper.infrastructure.login_status_repo import LoginStatusRepoImpl
-from atcoder_helper.infrastructure.task_config_repo import TaskConfigRepository
 from atcoder_helper.infrastructure.task_config_repo import TaskConfigRepositoryImpl
 
 T = TypeVar("T")
