+++ atcoder_helper/infrastructure/atcoder_helper_config_repo.py
@@ -1,42 +1,13 @@
 """AtcoderHelperConfigを永続化する層."""
 
 import os
-from typing import Protocol
 
 import yaml
 
+from atcoder_helper.application.repositories.errors import ParseError
+from atcoder_helper.application.repositories.errors import ReadError
+from atcoder_helper.application.repositories.errors import WriteError
 from atcoder_helper.entities.atcoder_helper_config import AtCoderHelperConfig
-from atcoder_helper.infrastructure.errors import ParseError
-from atcoder_helper.infrastructure.errors import ReadError
-from atcoder_helper.infrastructure.errors import WriteError
-
-
-class ConfigRepository(Protocol):
-    """AtCoderHelperConfigを永続化するプロトコル."""
-
-    def __init__(self, filename: str):
-        """__init__."""
-
-    def read(self) -> AtCoderHelperConfig:
-        """読み込みを行う.
-
-        Raises:
-            ReadError: 読み込みに失敗した
-            ParseError: パースに失敗した
-
-        Returns:
-            AtCoderHelperConfig: 読み込まれたAtcoderHelperConfig
-        """
-
-    def write(self, config: AtCoderHelperConfig) -> None:
-        """書き込みを行う.
-
-        Args:
-            config (AtCoderHelperConfig): _description_
-
-        Raises:
-            WriteError: 書き込みに失敗した
-        """
 
 
 class ConfigRepositoryImpl:
