+++ atcoder_helper/infrastructure/atcoder_logged_in_session_repo.py
@@ -1,35 +1,13 @@
 """atcoderからlogin済みのセッションを取得するrepository."""
 
 
-from typing import Protocol
-
 import requests
 from bs4 import BeautifulSoup
 
-from atcoder_helper.infrastructure.errors import ConnectionError
-from atcoder_helper.infrastructure.errors import LoginFailure
-from atcoder_helper.infrastructure.errors import ParseError
+from atcoder_helper.application.repositories.errors import ConnectionError
+from atcoder_helper.application.repositories.errors import LoginFailure
+from atcoder_helper.application.repositories.errors import ParseError
 from atcoder_helper.infrastructure.utils import AtCoderURLProvider
-
-
-class AtCoderLoggedInSessionRepository(Protocol):
-    """atcoder からlogin済みのセッションを取得するrepositoryのプロトコル."""
-
-    def read(self, username: str, password: str) -> requests.Session:
-        """atcoderにloginしたsessionを返す.入力したユーザーネームとパスワードは保存されず、代わりにセッションが保存される.
-
-        Args:
-            username (str): username
-            password (str): password
-
-        Raises:
-            ConnectionError: GETかPOSTに失敗
-            ParseError: パースに失敗
-            LoginFailure: ログイン失敗
-
-        Returns:
-            requests.Session: セッションを返す.
-        """
 
 
 class AtCoderLoggedInSessionRepositoryImpl:
