+++ atcoder_helper/infrastructure/atcoder_test_case_repo.py
@@ -2,35 +2,13 @@
 
 
 from typing import List
-from typing import Protocol
 
 import requests
 from bs4 import BeautifulSoup
 
+from atcoder_helper.application.repositories.errors import ParseError
 from atcoder_helper.entities.atcoder_test_case import AtcoderTestCase
-from atcoder_helper.infrastructure.errors import ParseError
 from atcoder_helper.infrastructure.utils import AtCoderURLProvider
-
-
-class AtCoderTestCaseRepository(Protocol):
-    """AtCoderからテストケースを取得するリポジトリのプロトコル."""
-
-    def fetch_test_cases(
-        self, session: requests.Session, contest: str, task: str
-    ) -> List[AtcoderTestCase]:
-        """テストケーススイートを取得する.
-
-        Args:
-            contest (str): コンテスト名
-            task (str): タスク名
-
-        Raises:
-            ConnectionError: GETに失敗
-            ParseError: Parseに失敗
-
-        Returns:
-            List[TestCase]: テストケーススイート
-        """
 
 
 class AtCoderTestCaseRepositoryImpl:
