+++ atcoder_helper/infrastructure/login_status_repo.py
@@ -1,25 +1,10 @@
 """loginしているかどうかを取得するrepository."""
 
-from typing import Protocol
 
 import requests
 
-from atcoder_helper.infrastructure.errors import ReadError
+from atcoder_helper.application.repositories.errors import ReadError
 from atcoder_helper.infrastructure.utils import AtCoderURLProvider
-
-
-class LoginStatusRepo(Protocol):
-    """login情報を取得するrepositoryのプロトコル."""
-
-    def is_logged_in(self, session: requests.Session) -> bool:
-        """loginしているかどうかを判定する.
-
-        Raises:
-            ReadError: タスクGETに失敗
-
-        Returns:
-            bool: loginしているか否か
-        """
 
 
 class LoginStatusRepoImpl:
