+++ tests/application/interactors/test_fetch_task.py
@@ -10,14 +10,14 @@ import pytest
 import requests
 
 from atcoder_helper.application.interactors.fetch_task import FetchTaskInteractor
+from atcoder_helper.application.repositories.errors import ConnectionError
+from atcoder_helper.application.repositories.errors import ParseError
+from atcoder_helper.application.repositories.errors import ReadError
+from atcoder_helper.application.repositories.errors import WriteError
 from atcoder_helper.application.usecases.errors import AtcoderAccessError
 from atcoder_helper.application.usecases.errors import ConfigAccessError
 from atcoder_helper.entities.atcoder_task_config import TaskConfig
 from atcoder_helper.entities.atcoder_test_case import AtcoderTestCase
-from atcoder_helper.infrastructure.errors import ConnectionError
-from atcoder_helper.infrastructure.errors import ParseError
-from atcoder_helper.infrastructure.errors import ReadError
-from atcoder_helper.infrastructure.errors import WriteError
 
 
 def _default_task_config(
